GCS_MAVLink: support 32 bit system IDs
Widen system ID handling throughout for MAVLink2.1 32 bit system IDs
(mavlink/rfcs#20):

 - MAV_SYSID, MAV_GCS_SYSID and MAV_GCS_SYSID_HI become AP_Int32 with
   parameter width conversion. Values must stay below 2^24 until the
   parameter transport moves away from float32
 - routing table entries, get_targets() and find_by_mavtype use the
   mavlink_msg_get_target_sysid/compid helpers so targets in the
   extended header are honoured
 - packet_overhead_chan() conservatively reserves the extended header
 - TSYN and MAVC log the full 32 bit source and target system IDs

+++ libraries/GCS_MAVLink/GCS_FTP.h
@@ -64,7 +64,7 @@ private:
         bool  burst_complete;
         uint8_t size;
         uint8_t session;
-        uint8_t sysid;
+        uint32_t sysid;
         uint8_t compid;
         uint8_t data[239];
     };
@@ -86,7 +86,7 @@ private:
         int16_t session_id;
         FTP_FILE_MODE mode; // work around AP_Filesystem not supporting file modes
         mavlink_channel_t chan;
-        uint8_t sysid;
+        uint32_t sysid;
         uint8_t compid;
 
         bool check_name_len(const Transaction &request);
